rename netcdf handler to match that of csv
